Update graded IMM-PDA task
From BB!
Edit 7. October: Added updated run_file and the estimationstatistics.
Also added a zip-file containing everything from the IMM-solution along with
the run file and data and should only miss pda.py and have the PDA-related TODOs
in IMM and EKF (should raise NotImplementedError if these methods are used as is).